feat: add kubectl wrapper and context protection (#6)
Add a wrapper around kubectl to support context protection. This can
help prevent accidentally running a protected

The context can be protected or unprotected explicitly which will be
saved to the state. `kubert context-protection protect` and `kubert
context-protection unprotect`

A default can also be set using a regex, the regex will be compared to
the context name and if it matches, the context is considered protected.
The default regex will be ignored if, for the respective context, a
protect/unprotect is explicitly present in the state. A
protect/unprotect can be deleted `kubert context-protection delete`,
this will cause the default to be used again.

When a context is protected, kubert will ask for confirmation when the
user wants to run a protected kubectl command.